requirements.txt
amqp==5.3.1
asgiref==3.9.0
billiard==4.2.1
celery==5.5.3
cfgv==3.4.0
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cron-descriptor==1.4.5
distlib==0.3.9
Django==5.2.4
django-celery-beat==2.8.1
django-cors-headers==4.7.0
django-environ==0.12.0
django-filter==25.1
django-redis==6.0.0
django-timezone-field==7.1
djangorestframework==3.16.0
djangorestframework-gis==1.2.0
drf-yasg==1.21.10
filelock==3.18.0
identify==2.6.12
inflection==0.5.1
kombu==5.5.4
nodeenv==1.9.1
packaging==25.0
platformdirs==4.3.8
pre_commit==4.2.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
python-crontab==3.2.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
redis==6.2.0
requests
rest-framework-simplejwt==0.0.2
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
vine==5.1.0
virtualenv==20.31.2
wcwidth==0.2.13
djangorestframework-simplejwt
pytest
pytest-django
pytz
factory-boy
